1st of December 2024 02/50*
This is my attempt to participate in the Advent of Code 2024 where this year I'm learning Zig.
Right now I'm writing a Scheme interpreter, called Plan-Z, in Zig as my main project.
In these commit messages I will be writing what I have learned, and any troubles and victories I've found,
a diary of sorts.

This was really the first time I've interacted with file IO in Zig. Not really the most fluent library I've seen,
but it does provide the basic infrastructure I need, and that's what's important in a systems language.
For instance I had to define memory maximums when reading lines of text, something that I'm not used to when programming,
even in C!

What stumped me the most was the most simple thing, parsing two integers in a single line!
The solution I found was to split the line into two slices, I did it in a manual way, by checking with std.ascii.isDigit
and finding the where each number can be, and each slice is parsed to an integer, this was another lesson in the importance
of slices in Zig.

This solution takes in average 0.004 seconds to run (in my machine.)