6th of December 2024 - Optimize the map structure
I changed the implementation of the map from a slice of slices to a
simple array. There was some improvement in the order of 20% speedup.

However this is still not the best algorithm to solve, because this
solution still took around 500x longer to run than the previous days.

Sure it runs in a bit over 200 ms, but coming from 500 μs it's a steep
difference.